Draft parsing JSON expressions from the front end
I decided it would be faster to parse the expressions
in the front end then serialize them to JSON then parse
the JSON again in the backend. Parsing expressions seems
outside the scope of computer algebra, more of a front end
task.

Also added the Variable expression type to the backend.